Turn on -Wconversion and -Wsign-conversion for our code
Both flags, because clang's -Wconversion includes sign changes and GCC's
leaves them to -Wsign-conversion. A clean clang build under -Werror and a
GCC 16 syntax pass over every one of our translation units now report
nothing. The findings were fixed by changing types, not by casting:

- yaml-cpp 0.9.0's public headers passed a signed stream precision to a
  size_t parameter. Patched with upstream's own post-0.9.0 change.
- DBC codegen: multiplexor_group_indexes and start_mux_group_index are the
  multiplexor's Raw type, emitted after the signal traits. The TC8 mock and
  the API test set the mux member through from_raw() instead of assigning
  a raw index to a physical value.
- OSM way ids go from the graph's int64 to the wire's UInt64 through
  map_wire::osmWayIdOf(), with a test.
- Iterator-plus-size_t arithmetic became std::span first()/subspan()
  (tlv8, the AirPlay media stream, a pub_sub test).
- Elapsed times use std::chrono::duration<double> rather than an integer
  count divided by 1000.0 (road_graph contraction, map_build, the test
  data publisher).
- Byte assembly works in unsigned widths: the MoTeC gateway CRC in uint32_t
  with one narrowing at the end, an MFi big-endian length helper, and the
  MFi certificate dates no longer turn a negative BIO length into a huge
  size_t.
- Widgets do geometry in qreal and index in size_t: centred scaling, the
  190E cluster, speedometer and tachometer, the CDL3 arc tables, sparkline,
  segment and value readouts. The speedometer's odometer is uint32_t like
  its config. value_readout converts Qt::Alignment to drawText's int flags
  once.
- Integer temperatures from the TC8 are converted to the schema's Float32
  explicitly; every value of the 16-bit field is exact.
- Tests: counting with if/++ instead of adding a ternary int to size_t,
  typed loops, and map_match now checks the Result of writing its graphs,
  which GCC reported as ignored.

AGENTS.md and the developer docs say the flags are on and how to fix a
finding.

+++ libs/airplay/test_info_plist.cpp
@@ -113,7 +113,8 @@ int main()
         {
             const double pixel_ratio = static_cast<double>(config.width) / config.height;
             const double physical_ratio =
-                static_cast<double>(width_physical->asInteger()) / height_physical->asInteger();
+                static_cast<double>(width_physical->asInteger()) /
+                static_cast<double>(height_physical->asInteger());
             expect(std::abs(pixel_ratio - physical_ratio) < 0.05,
                    "physical size keeps the pixel aspect ratio");
         }
